[Refactor] #430 각 TextField 별로 파일 분리 및 이름 수정
